moved attention related files into attention subfolder
