+++ src/models/huggingface/mpt.h
@@ -5,7 +5,7 @@
 #include <optional>
 
 #include "layers/activation.h"
-#include "layers/attention_alibi.h"
+#include "layers/attention/attention_alibi.h"
 #include "layers/embedding.h"
 #include "layers/linear.h"
 #include "layers/normalization.h"
